requirements.txt
openai>=2.0.0
python-dotenv>=1.0.0
langgraph>=1.0.0
chromadb>=0.5.0
aiogram>=3.0.0
pytest>=9.0.0
